+++ Relentless/Relentless/Game Logic/GameParameters/GameHostParameters.swift
@@ -76,8 +76,8 @@ class GameHostParameters: GameParameters {
         return probOfEventExpression?(varDict) ?? defaultValue
     }
 
-    func probOfSelectingAssembledItem(numberOfPlayers: Int) -> Float {
-        let defaultValue = 1 / Float(numberOfPlayers)
+    func probOfSelectingAssembledItem(numberOfPlayers: Int) -> Double {
+        let defaultValue = 1 / Double(numberOfPlayers)
         return probOfSelectingAssembledItemExpression?(varDict) ?? defaultValue
     }
